+++ src/ISiteDesignSchemaConfiguration.ts
@@ -1,0 +1,4 @@
+export interface ISiteDesignSchemaConfiguration {
+  schemaUrl: string
+  allowTelemetry: boolean,
+}