+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "sitedesign-schema" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [1.1.0] - 2020-07-14
+### Added
+- Added configuration for schema url & preferences
+- Added anonymous telemetry for improved exception tracking
 
-## [Unreleased]
-
+## [1.0.0] - 2020-07-13
+### Added
 - Initial release